Actually finishes CameraServer shared

+++ cameraserver/src/main/java/edu/wpi/first/wpilibj/sharedinterfaces/RobotMainThreadIdFunc.java
@@ -1,5 +1,5 @@
 package edu.wpi.first.wpilibj.sharedinterfaces;
 
 public interface RobotMainThreadIdFunc {
-  Integer getMainThreadId();
+  Long getMainThreadId();
 }
